refactor: DayOfWeek (WIP)
