+++ r4/src/main/java/io/github/jy95/fds/r4/utils/maps/TimingRepeatTranslatorsMapR4.java
@@ -7,6 +7,7 @@ import java.util.function.Supplier;
 import org.hl7.fhir.r4.model.Timing.TimingRepeatComponent;
 
 import io.github.jy95.fds.common.functions.TranslationService;
+import io.github.jy95.fds.common.translators.timing.repeat.DayOfWeek;
 import io.github.jy95.fds.common.translators.timing.repeat.TimeOfDay;
 import io.github.jy95.fds.common.types.AbstractTranslatorsMap;
 import io.github.jy95.fds.common.types.DisplayOrder;
@@ -48,7 +49,11 @@ public class TimingRepeatTranslatorsMapR4 extends AbstractTranslatorsMap<FDSConf
         suppliers.put(DisplayOrder.DURATION_DURATION_MAX, () -> new DurationDurationMaxR4(translationService));
         suppliers.put(DisplayOrder.FREQUENCY_FREQUENCY_MAX, () -> new FrequencyFrequencyMaxR4(translationService));
         suppliers.put(DisplayOrder.PERIOD_PERIOD_MAX, () -> new PeriodPeriodMaxR4(translationService));
-        suppliers.put(DisplayOrder.DAY_OF_WEEK, () -> new DayOfWeekR4(translationService));
+        suppliers.put(DisplayOrder.DAY_OF_WEEK, () -> new DayOfWeek<>(
+                translationService,
+                TimingRepeatComponent::getDayOfWeek,
+                TimingRepeatComponent::hasDayOfWeek
+        ));
         suppliers.put(DisplayOrder.TIME_OF_DAY, () -> new TimeOfDay<>(
                 translationService,
                 TimingRepeatComponent::getTimeOfDay,
